requirements.txt
flask
firecrawl
groq
google-generativeai
markdown
python-dotenv